go.exp/locale: further implementation of locale API: - Added parser for IANA Language Subtag Registry.  Some of the code is replaced   to use this data source where it is more complete than the CLDR data. - Parser now supports identifying about 8000 languages (instead of 600+ before). - Language, Script, and Region type are implemented. - Canonicalization implemented for:         - script suppression, as defined by BCP 47/ IANA registry.         - mapping of deprecated languages, including 3-letter deprecated codes.         - mapping of dominant languages to their macro language, as defined by CLDR. - Fixed bug where the 3-letter ISO code for sh was scr, instead of hbs.
LangIDs are now split into IDs for which we look up data in a lookup table
(includes all 2-letter codes) and others.  Only the former are mapped to a
consecutive ranges of ids.  The other langIDs to be generated from
the 3-letter languages code and their existence is represented by a bit vector.

The API now supports representing of deprecated language tags. They will be mapped
for all typical use cases, but if the user really wants he can investigate
the language that was originally specified.

R=r, mpvl
CC=golang-dev
https://codereview.appspot.com/8594044