exp/ssa: support for calls f(g()) where g has multiple return values (MRVs).
The new logic is much simpler:
(1) emit code for all actual parameter expressions and make a
    list of SSA values.
(2) in the "chained MRV" case, we flatten out the tuple from
    g() into this list.
(3) we apply assignability conversions of the prefix
    corresponding to normal formal parameters.
(4) we apply assignability conversions and build a slice for
    the suffix corresponding to the variadic formal parameter.

Ellipsis calls are handled as a simple special case.

Built-ins are handled using exactly the same code, but we must
first infer a types.Signature for the callsite.

Tested in testdata/mrvchain.go and $GOROOT/test/simassign.go.

Also in this CL:

- fix: apply assignability conversions when returning a single
  multi-valued expression in an MRV function.  Added test.
- fix: ensure that createSyntacticParams creates
  Function.Params for anonymous parameters.
- ssa.Select now uses the empty interface type to represent
  the recv parameter.  I shoulda done this earlier...
- ssa.Function now has an AnonFuncs field.
  (Required for converting a token.Pos to its enclosing
  *ssa.Function)
- ssa.Package.Pos is gone; it was bogus.
- Set the Package.Init.Pos correctly: to the lexically first
  'func init()' decl, or zero if implicit.
- Documented some doubts about the spec of ChangeInterface,
  TypeAssert.

R=gri, minux.ma
CC=golang-dev
https://codereview.appspot.com/8569046